split BillingProductAddon to its own file
